+++ CHANGELOG.rst
@@ -17,6 +17,7 @@ Changed
 - Optimized some functions and comparisons according to the recommended Python optimization instructions 
 - ``ContactMap.match`` does __not__ modifiy ``other`` by default anymore. Specify ``match_other=True`` as kwarg!
 - ``ContactMap.calculate_kernel_density`` renamed to ``ContactMap.calculate_contact_density`` 
+- ``ContactDensityFigure`` draws domain boundary lines instead of symbols
 
 [0.8.3]
 -------